major: switch from boltdb/bolt to coreos/bbolt
The original BoltDB project has been archived and is no longer being
maintained. CoreOS has forked it as github.com/coreos/bbolt in order to
provide a maintained version of BoltDB with backwards compatability.

While all of the tests pass without any changes, this doe replace this
projects one and only dependency so it should probably be considered a
semver-major change.